Don't try to dump empty models specified by MODELS.
This fixes a bug that caused the seed file to be appended to even
when the `append` parameter is set to false. The root of the bug
is that models specified the `model[s]` option were not being checked
for emptiness. This changeset causes such models to be checked
for emptiness. Fixes #65.