Add the ability to add options for activerecord_import
There are two options, one already in the gem, and one would be added. 
The first is to avoid validation (which is good, considering you just need to load the data that previously exist in a similar data). 
The second is to avoid the serialization done by activerecord_import on serialized fields, that doesn't seem to work well in all cases. (works only for hashes?)